More efficient tuple basis conversions
